deep_research_app/requirements.txt
streamlit>=1.28.0
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
firecrawl-py>=0.0.16
reportlab>=4.0.0
